+++ src-tauri/src/manager/downloader/commands.rs
@@ -11,7 +11,7 @@ use itertools::Itertools;
 
 #[tauri::command]
 pub async fn install_mod(mod_ref: ModRef, app: tauri::AppHandle) -> Result<()> {
-    super::install_with_deps(&mod_ref, InstallOptions::default(), &app).await?;
+    super::install_with_deps(&[(mod_ref, true)], InstallOptions::default(), &app).await?;
 
     Ok(())
 }