+++ src/lib/modlist/ModListItem.svelte
@@ -16,7 +16,6 @@
 		? 'bg-slate-700'
 		: 'hover:bg-slate-700 border-opacity-0'}"
 	data-uuid={mod.uuid}
-	data-name={mod.name}
 	on:click
 	on:dragstart
 	on:dragover